Web-site - added checking image for null before deleting in about page

+++ my-house24-admin/src/main/java/com/example/myhouse24admin/serviceImpl/AboutPageServiceImpl.java
@@ -152,8 +152,13 @@ public class AboutPageServiceImpl implements AboutPageService {
         if(image.isEmpty()){
             return aboutPage.getDirectorImage();
         } else {
-            uploadFileUtil.deleteFile(aboutPage.getDirectorImage());
+            if(imageNotNull(aboutPage)) {
+                uploadFileUtil.deleteFile(aboutPage.getDirectorImage());
+            }
             return uploadFileUtil.saveMultipartFile(image);
         }
+    }
+    private boolean imageNotNull(AboutPage aboutPage){
+        return aboutPage.getDirectorImage() != null;
     }
 }
